refactor: move compressor to modules (#2)
