+++ modules/compressor/zlib/zlib.go
@@ -1,4 +1,4 @@
-// Copyright 2023-2023 EMQ Technologies Co., Ltd.
+// Copyright 2023-2024 EMQ Technologies Co., Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,60 +12,60 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gzip
+package zlib
 
 import (
 	"bytes"
 	"fmt"
 	"io"
 
-	"github.com/klauspost/compress/gzip"
+	"github.com/klauspost/compress/zlib"
 
 	"github.com/lf-edge/ekuiper/v2/internal/conf"
 )
 
-func NewGzipCompressor() (*gzipCompressor, error) {
-	return &gzipCompressor{
-		writer: gzip.NewWriter(nil),
+func NewZlibCompressor() (*zlibCompressor, error) {
+	return &zlibCompressor{
+		writer: zlib.NewWriter(nil),
 	}, nil
 }
 
-type gzipCompressor struct {
-	writer *gzip.Writer
+type zlibCompressor struct {
+	writer *zlib.Writer
 	buffer bytes.Buffer
 }
 
-func (g *gzipCompressor) Compress(data []byte) ([]byte, error) {
-	g.buffer.Reset()
-	g.writer.Reset(&g.buffer)
-	_, err := g.writer.Write(data)
+func (z *zlibCompressor) Compress(data []byte) ([]byte, error) {
+	z.buffer.Reset()
+	z.writer.Reset(&z.buffer)
+	_, err := z.writer.Write(data)
 	if err != nil {
 		return nil, err
 	}
-	err = g.writer.Close()
+	err = z.writer.Close()
 	if err != nil {
 		return nil, err
 	}
-	return g.buffer.Bytes(), nil
+	return z.buffer.Bytes(), nil
 }
 
-func NewGzipDecompressor() (*gzipDecompressor, error) {
-	return &gzipDecompressor{}, nil
+func NewZlibDecompressor() (*zlibDecompressor, error) {
+	return &zlibDecompressor{}, nil
 }
 
-type gzipDecompressor struct {
-	reader *gzip.Reader
+type zlibDecompressor struct {
+	reader io.ReadCloser
 }
 
-func (z *gzipDecompressor) Decompress(data []byte) ([]byte, error) {
+func (z *zlibDecompressor) Decompress(data []byte) ([]byte, error) {
 	if z.reader == nil {
-		r, err := gzip.NewReader(bytes.NewReader(data))
+		r, err := zlib.NewReader(bytes.NewReader(data))
 		if err != nil {
 			return nil, fmt.Errorf("failed to decompress: %v", err)
 		}
 		z.reader = r
 	} else {
-		err := z.reader.Reset(bytes.NewReader(data))
+		err := z.reader.(zlib.Resetter).Reset(bytes.NewReader(data), nil)
 		if err != nil {
 			return nil, fmt.Errorf("failed to decompress: %v", err)
 		}
@@ -73,16 +73,8 @@ func (z *gzipDecompressor) Decompress(data []byte) ([]byte, error) {
 	defer func() {
 		err := z.reader.Close()
 		if err != nil {
-			conf.Log.Warnf("failed to close gzip decompressor: %v", err)
+			conf.Log.Warnf("failed to close zlib decompressor: %v", err)
 		}
 	}()
 	return io.ReadAll(z.reader)
-}
-
-func NewReader(r io.Reader) (io.ReadCloser, error) {
-	return gzip.NewReader(r)
-}
-
-func NewWriter(w io.Writer) (io.Writer, error) {
-	return gzip.NewWriter(w), nil
 }
